Dynamically import classes
